+++ frontend/src/pages/FeedVagas/FeedVagas.vue
@@ -97,7 +97,7 @@
             :titulo="anuncio.titulo"
             :area="anuncio.areaEmprego.nome"
             :descricao="anuncio.descricao"
-            :salario="anuncio.salario"
+            :salario="formataSalario(anuncio.salario)"
             :foto="anuncio.createdByUser.foto"
           />
         </div>
@@ -195,6 +195,18 @@ const toggleFilterApplied = (id:number) => {
 
 const applyFilters = (filters:any) => {
   filtersById.value = filters.map((elem: any) => (elem.id))
+}
+
+const formataSalario = (value:String) => {
+  const valueConvertido = parseFloat(value.toString())
+  const formattedValue = valueConvertido.toLocaleString('pt-BR', {
+    style: 'currency',
+    currency: 'BRL'
+  })
+  if (valueConvertido !== 0.00) {
+    return formattedValue
+  }
+  return ''
 }
 
 </script>
